Use robolectric for clickHandler tests, removing PowerMock.
